Xiami: add as lyrics provider (read notes)
This is the first, as-straightforward-as-possible, lyrics
implementation.
Surely it will not work as intended, also because there's a whole
missing part about getting lyrics properly (but I couldn't manage to get
it working properly after hours of tinkering so I am happy with this
one instead).
It might be removed if it doesn't play well with MPRISLyrics.

PS: Contributions are very welcome on this provider.